change the data type of final[2]
int final[2] --> uint16_t final[2]

+++ WINC1500_HTTP_DOWNLOADER_EXAMPLE1/src/WifiHandlerThread/WifiHandler.h
@@ -97,8 +97,10 @@ struct TestPacket {
 };
 
 struct NauPacket {
-	uint32_t nau_i;
-	uint32_t nau_f;
+	uint16_t nau_i;
+	uint16_t nau_f;
+	//uint32_t nau_i;
+	//uint32_t nau_f;
 };
 
 /* Max size of UART buffer. */
